organize statusline integration & split into multiple files
