+++ lua/base46/integrations/defaults.lua
@@ -245,7 +245,6 @@ local defaults = {
 
 -- merge statusilne & hl_add tables!
 local merge_tb = require("base46").merge_tb
-defaults = merge_tb(defaults, require("base46").load_integrationTB "statusline")
 
 local user_new_highlights = require("nvconfig").ui.hl_add
